+++ include/XTlvObject.h
@@ -27,7 +27,7 @@ public:
 		BLOCK_TOO_SHORT
 	} XType;
 	static const char *XTypeString[];
-	explicit XTlvObject(XType e, uint16_t id = TLV_TYPE_INVALID,
+	explicit XTlvObject(XType e, uint32_t id = TLV_TYPE_INVALID,
 			const std::string &name = "?") throw();
 	virtual ~XTlvObject() throw() {}
 	inline XType xtype() const throw() { return _e; }
