+++ test/unittest/TLVUIntCreatorTestSuite.cpp
@@ -18,7 +18,7 @@ void TLVUIntCreatorTestSuite::testUInt16()
 {
 	CTlvUint16 u16(5566);
 	ITlvBlock *blk = u16.toTLVBlock();
-	CPPUNIT_ASSERT_EQUAL((unsigned short)2, blk->length());
+	CPPUNIT_ASSERT_EQUAL((uint32_t)2, blk->length());
 
 	CTlvUint16 *obj = dynamic_cast<CTlvUint16 *>(CTlvUint16Creator().create(*blk));
 	CPPUNIT_ASSERT_EQUAL(u16.value(), obj->value());
@@ -31,7 +31,7 @@ void TLVUIntCreatorTestSuite::testUInt32()
 {
 	CTlvUint32 u32(7788);
 	ITlvBlock *blk = u32.toTLVBlock();
-	CPPUNIT_ASSERT_EQUAL((unsigned short)4, blk->length());
+	CPPUNIT_ASSERT_EQUAL((uint32_t)4, blk->length());
 
 	CTlvUint32 *obj = dynamic_cast<CTlvUint32 *>(CTlvUint32Creator().create(*blk));
 	CPPUNIT_ASSERT_EQUAL(u32.value(), obj->value());
